[#20241018] Separate domain logic into separate files
- Splits domain-related classes into individual files for better
modularity
- Improves maintainability and clarity of the domain logic
